# source/gismolab-swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "GISMoLab API"
    },
    "paths": {
...
        "/evchargers": {
            "get": {
                "summary": "Get EV charger device list",
                "description": "Returns a list of available EV charger devices.",
                "responses": {
                    "200": {
                        "description": "list of EV charger names",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
...
        "/device/{device_id}/start": {
            "get": {
                "summary": "Start device command handling",
                "description": "",
                "responses": {
                    "200": {
                        "description": "Confirmation",
                        "content": "application/json",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "description": "User access token"
                                        }
                                    },
                                    "description": "Data payload"
                                },
                                "status": {
                                    "type": "string",
                                    "description": "Response status"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "content": "application/json",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error message"
                                },
                                "status": {
                                    "type": "string",
                                    "description": "Response status"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Device is busy",
                        "content": "application/json",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error message"
                                },
                                "status": {
                                    "type": "string",
                                    "description": "Response status"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "device_id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Device identifier"
                    }
                ]
            }
        },
        "/device/{device_id}/stop": {
            "get": {
                "summary": "Stop device command handling",
                "description": "",
                "responses": {
                    "200": {
                        "description": "Confirmation",
                        "content": "application/json",
                        "schema": {
                            "type": "dict"
                        }
                    },
                    "408": {
                        "description": "Device not found",
                        "content": "application/json",
                        "schema": {
                            "type": "dict"
                        }
                    },
                    "410": {
                        "description": "Device not waiting",
                        "content": "application/json",
                        "schema": {
                            "type": "dict"
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "device_id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Device identifier"
                    },
                    {
                        "in": "query",
                        "name": "token",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Access token"
                    }
                ]
            }
        },
        "/device/{device_id}/recv": {
            "get": {
                "summary": "Receive device command",
                "description": "",
                "responses": {
                    "200": {
                        "description": "Device command",
                        "content": "application/json",
                        "schema": {
                            "type": "dict"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": "application/json",
                        "schema": {
                            "type": "dict"
                        }
                    },
                    "405": {
                        "description": "Not allowed",
                        "content": "application/json",
                        "schema": {
                            "type": "dict"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "content": "application/json",
                        "schema": {
                            "type": "dict"
                        }
                    },
                    "408": {
                        "description": "Receive timeout",
                        "content": "application/json",
                        "schema": {
                            "type": "dict"
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "device_id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Device identifier"
                    },
                    {
                        "in": "query",
                        "name": "token",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Access token"
                    },
                    {
                        "in": "query",
                        "name": "timeout",
                        "schema": {
                            "type": "int"
                        },
                        "description": "Receive timeout in seconds"
                    }
                ]
            }
        },
...
    },
...
    "servers": {
        "url": "https://0.0.0.0:5001/"
    }
}